DBInterface fix

+++ src/main/java/project/models/Task.java
@@ -6,21 +6,21 @@ import java.util.Objects;
  * @author Ilia Moskalenko
  */
 public class Task{
-    private int id;
+    private long id;
     private String title;
     private String fullTask;
     private boolean solved;
 
 
 
-    public Task(int id, String title, String fullTask) {
+    public Task(long id, String title, String fullTask) {
         this.id = id;
         this.title = title;
         this.fullTask = fullTask;
         solved = false;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
